Add files via upload
In this mini project/assignment, we have analyzed the traces of 4 programs by simulating a machine with no cache. Followed by that, we have analyzed our results on that cache less simulated system and have analyzed the Cumulative Density Function of the misses. From which we have gained some insights and have introduced a cache in our simulation model, trying to observe how by this new change we are benefited. We have received 100% marks for this experiment of ours. :)